@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.15",
+    "@angular/animations": "8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
     "@angular/flex-layout": "7.0.0-beta.23",
-    "@angular/forms": "7.2.15",
+    "@angular/forms": "8.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@auth0/angular-jwt": "^4.2.0",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -46,19 +46,19 @@
     "primeng": "^7.1.3",
     "rxjs": "^6.6.3",
     "tslib": "^2.0.1",
-    "zone.js": "^0.8.29",
+    "zone.js": "~0.9.1",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.10",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
     "@types/jasmine": "^2.8.17",
     "@types/node": "^9.6.58",
     "babel-preset-env": "^1.7.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "intl": "^1.2.5",
     "jasmine-core": "^2.99.1",
     "jasmine-spec-reporter": "4.2.1",
@@ -75,7 +75,7 @@
     "puppeteer": "^1.20.0",
     "ts-node": "^4.1.0",
     "tslint": "^5.20.1",
-    "typescript": "3.2.4",
+    "typescript": "3.5.3",
     "webpack": "^4.44.1"
   }
 }
