@@ -42,10 +42,10 @@ export function noParameterTypeInControlValueValidator(): ValidatorFn {
 export class AppVersionCreateWizardComponent extends BaseComponent implements OnInit {
 
 
-    @ViewChild(ModalComponent)
+    @ViewChild(ModalComponent, { static: true })
     public modal: ModalComponent;
 
-    @ViewChild('tagsMultiSelect')
+    @ViewChild('tagsMultiSelect', { static: false })
     public tagsMultiSelect: MultiSelect;
 
     public app: Application;
