@@ -8,7 +8,7 @@ import {ModalComponent} from '../modal.component';
 })
 export class ModalProvideSshKeyComponent implements OnInit {
 
-  @ViewChild(ModalComponent)
+  @ViewChild(ModalComponent, { static: true })
   public readonly modal: ModalComponent;
 
   constructor() { }
