@@ -9,7 +9,7 @@ import {ModalComponent} from '../modal.component';
 })
 export class ModalTestInstanceComponent implements OnInit {
 
-  @ViewChild(ModalComponent)
+  @ViewChild(ModalComponent, { static: true })
   public readonly modal: ModalComponent;
 
   constructor() { }
